# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для биллинга",
        "description": "Этот API предоставляет систему биллинга, в которой пользователи могут зарегистрироваться, проверять свой баланс, выполнять транзакции (пополнение/снятие) и просматривать историю транзакций. Все операции проходят проверку, исключая возможность отрицательного баланса.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер для разработки API биллинга."
        }
    ],
    "paths": {
...
        "/api/user/info": {
            "get": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Информация о пользователе",
                "description": "Этот метод позволяет пользователю получить информацию о своей учетной записи, включая баланс и другие данные.",
                "operationId": "cdf71a463c6c18c5370d453b2fbd0437",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "content": {
                                            "description": "Информация о пользователе",
                                            "properties": {
                                                "id": {
                                                    "description": "ID пользователя",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Имя пользователя",
                                                    "type": "string",
                                                    "example": "Dr. Savion Stroman"
                                                },
                                                "email": {
                                                    "description": "Email пользователя",
                                                    "type": "string",
                                                    "example": "mr@test.com"
                                                },
                                                "created_at": {
                                                    "description": "Дата и время создания учетной записи",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-01-31T07:13:14.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "Дата и время последнего обновления учетной записи",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-01-31T07:13:14.000000Z"
                                                },
                                                "balance": {
                                                    "description": "Баланс пользователя",
                                                    "type": "string",
                                                    "example": "170.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке аутентификации",
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Используйте Bearer Token для аутентификации. Все защищенные маршруты требуют действительного токена после входа пользователя в систему.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}